fix(ui): Fix the formatting of time remaining.
The bug or faulty display of time was caused by how I round up the time.
When there's 1m 59.995s remaining, it got rounded up to 1m 60s, thus the
faulty display. After fixing, the current time is rounded to the exact
second, so is remaining time, so instead of 1m 59.995s, we have 2m 0s,
which is perfect for displaying the seconds.